deployment任务模版更新

+++ caesar-service/src/main/java/com/baiyi/caesar/service/jenkins/impl/CsCiJobServiceImpl.java
@@ -26,14 +26,14 @@ public class CsCiJobServiceImpl implements CsCiJobService {
 
     @Override
     public DataTable<CsCiJob> queryCsCiJobByParam(CiJobParam.CiJobPageQuery pageQuery) {
-        Page page = PageHelper.startPage(pageQuery.getPage(), pageQuery.getLength().intValue());
+        Page page = PageHelper.startPage(pageQuery.getPage(), pageQuery.getLength());
         List<CsCiJob> list = csCiJobMapper.queryCsCiJobByParam(pageQuery);
         return new DataTable<>(list, page.getTotal());
     }
 
     @Override
     public DataTable<CsCiJob> queryCsCiJobByParam(CiJobParam.CiJobTplPageQuery pageQuery) {
-        Page page = PageHelper.startPage(pageQuery.getPage(), pageQuery.getLength().intValue());
+        Page page = PageHelper.startPage(pageQuery.getPage(), pageQuery.getLength());
         List<CsCiJob> list = csCiJobMapper.queryCsCiJobTplByParam(pageQuery);
         return new DataTable<>(list, page.getTotal());
     }
